fix api (exclude cart)

+++ lib_artiexh-model/src/main/java/com/artiexh/model/domain/Product.java
@@ -13,21 +13,21 @@ import java.util.Set;
 @AllArgsConstructor
 @SuperBuilder
 public class Product {
-    private Long id;
-    private Artist owner;
-    private ProductStatus status;
-    private String name;
-    private Money price;
-    private String thumbnailUrl;
-    private ProductCategory category;
-    private String description;
-    private ProductType type;
-    private Long remainingQuantity;
-    private Instant publishDatetime;
-    private Long maxItemsPerOrder;
-    private DeliveryType deliveryType;
-    private Float averageRate;
-    private Set<Byte> paymentMethods;
-    private Set<ProductTag> tags;
-    private Set<ProductAttach> attaches;
+	private Long id;
+	private Artist owner;
+	private ProductStatus status;
+	private String name;
+	private Money price;
+	private String thumbnailUrl;
+	private ProductCategory category;
+	private String description;
+	private ProductType type;
+	private Integer remainingQuantity;
+	private Instant publishDatetime;
+	private Long maxItemsPerOrder;
+	private DeliveryType deliveryType;
+	private Float averageRate;
+	private Set<Byte> paymentMethods;
+	private Set<ProductTag> tags;
+	private Set<ProductAttach> attaches;
 }
